Feature/obj (#20)
* Added mesh primitive types
* Added function to toggle wireframe mode
* Add Noise lib
* Added Height shader.
* Add Triangle grid to MeshManager
* Cache shaders
* Add third party code (for building offline)
* Add assimp
* Load objs, test with sponza model

+++ examples/cube/CubeOverlayComponent.cpp
@@ -2,6 +2,7 @@
 #include <FIRE/Message.h>
 #include <FIRE/TextOverlay.h>
 #include <FIRE/TextRenderer.h>
+#include <iomanip>
 #include <sstream>
 namespace examples
 {
@@ -14,7 +15,8 @@ CubeOverlayComponent::CubeOverlayComponent(FIRE::TextRenderer& renderer)
 void CubeOverlayComponent::DoUpdate(double deltaTime, FIRE::SceneObject&, FIRE::Scene&)
 {
     std::stringstream ss;
-    ss << (1.0 / deltaTime);
+    ss << static_cast<int>(1.0 / deltaTime);
+
     renderer.Submit(FIRE::TextOverlay(0, ss.str(), 0.02f, 0.02f, 0.5f));
 }
